+++ menuentryswapperextended/src/main/java/net/runelite/client/plugins/menuentryswapperextended/util/GamesNecklaceMode.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2018, https://openosrs.com
- * Copyright (c) 2018, Kyle <https://github.com/kyleeld>
+ * Copyright (c) 2019, Alan Baumgartner <https://github.com/alanbaumgartner>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -23,18 +22,20 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.menuentryswapperextended.util.teleportmode;
+package net.runelite.client.plugins.menuentryswapperextended.util;
 
-public enum NecklaceOfPassageMode
+public enum GamesNecklaceMode
 {
 	OFF("Off"),
-	WIZARDS_TOWER("Wizards' Tower"),
-	THE_OUTPOST("The Outpost"),
-	EAGLES_EYRIE("Eagles' Eyrie");
+	BURTHORPE("Burthorpe"),
+	BARBARIAN_OUTPOST("Barbarian Outpost"),
+	CORPOREAL_BEAST("Corporeal Beast"),
+	TEARS_OF_GUTHIX("Tears of Guthix"),
+	WINTER("Wintertodt Camp");
 
 	private final String name;
 
-	NecklaceOfPassageMode(String name)
+	GamesNecklaceMode(String name)
 	{
 		this.name = name;
 	}
