+++ menuentryswapperextended/src/main/java/net/runelite/client/plugins/menuentryswapperextended/util/GloryMode.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2018, https://openosrs.com
- * Copyright (c) 2018, Kyle <https://github.com/kyleeld>
+ * Copyright (c) 2019, Alan Baumgartner <https://github.com/alanbaumgartner>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -23,18 +22,19 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.menuentryswapperextended.util.teleportmode;
+package net.runelite.client.plugins.menuentryswapperextended.util;
 
-public enum RingOfWealthMode
+public enum GloryMode
 {
-	MISCELLANIA("Miscellania"),
-	GRAND_EXCHANGE("Grand Exchange"),
-	FALADOR("Falador"),
-	DONDAKAN("Dondakan");
+	OFF("Off"),
+	EDGEVILLE("Edgeville"),
+	KARAMJA("Karamja"),
+	DRAYNOR_VILLAGE("Draynor Village"),
+	AL_KHARID("Al Kharid");
 
 	private final String name;
 
-	RingOfWealthMode(String name)
+	GloryMode(String name)
 	{
 		this.name = name;
 	}
