+++ menuentryswapperextended/src/main/java/net/runelite/client/plugins/menuentryswapperextended/util/XericsTalismanMode.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2019, Owain van Brakel <https://github.com/Owain94>
+ * Copyright (c) 2018, https://openosrs.com
+ * Copyright (c) 2018, Kyle <https://github.com/kyleeld>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -24,54 +25,24 @@
  */
 package net.runelite.client.plugins.menuentryswapperextended.util;
 
-import com.google.common.base.Splitter;
-import java.util.List;
-import javax.inject.Singleton;
-
-@Singleton
-public class PrioParse
+public enum XericsTalismanMode
 {
-	public static boolean parse(String value)
+	XERICS_LOOKOUT("Xeric's Look-out"),
+	XERICS_GLADE("Xeric's Glade"),
+	XERICS_INFERNO("Xeric's Inferno"),
+	XERICS_HEART("Xeric's Heart"),
+	XERICS_HONOUR("Xeric's Honour");
+
+	private final String name;
+
+	XericsTalismanMode(String name)
 	{
-		if (value.equals(""))
-		{
-			return true;
-		}
+		this.name = name;
+	}
 
-		try
-		{
-			final StringBuilder sb = new StringBuilder();
-
-			for (String str : value.split("\n"))
-			{
-				if (!str.startsWith("//"))
-				{
-					sb.append(str).append("\n");
-				}
-			}
-
-			final Splitter NEWLINE_SPLITTER = Splitter
-				.on("\n")
-				.omitEmptyStrings()
-				.trimResults();
-
-			final List<String> tmp = NEWLINE_SPLITTER.splitToList(sb);
-
-			for (String s : tmp)
-			{
-				final String[] strings = s.split(",");
-
-				if (strings.length <= 1)
-				{
-					return false;
-				}
-			}
-
-			return tmp.size() > 0;
-		}
-		catch (Exception ex)
-		{
-			return false;
-		}
+	@Override
+	public String toString()
+	{
+		return name;
 	}
 }
